requirements.txt
anthropic
resend
python-dotenv
flask
flask-cors
